backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.1
python-dotenv==1.1.1
sentence-transformers==5.1.0
openai==1.68.2
